refactor: create util folder, move nap_sym to napi/sym
