+++ Cargo.toml
@@ -5,7 +5,7 @@ resolver = "2"
 members = [
   "bench_util",
   "cli",
-  "cli/napi_sym",
+  "cli/napi/sym",
   "core",
   "ops",
   "runtime",
@@ -47,7 +47,7 @@ deno_core = { version = "0.161.0", path = "./core" }
 deno_ops = { version = "0.39.0", path = "./ops" }
 serde_v8 = { version = "0.72.0", path = "./serde_v8" }
 deno_runtime = { version = "0.87.0", path = "./runtime" }
-napi_sym = { version = "0.9.0", path = "./cli/napi_sym" }
+napi_sym = { version = "0.9.0", path = "./cli/napi/sym" }
 deno_bench_util = { version = "0.73.0", path = "./bench_util" }
 test_util = { path = "./test_util" }
