+++ cli/cache/disk_cache.rs
@@ -1,7 +1,7 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 
-use crate::fs_util;
 use crate::http_cache::url_to_filename;
+use crate::util::fs::atomic_write_file;
 
 use deno_core::url::Host;
 use deno_core::url::Url;
@@ -144,7 +144,7 @@ impl DiskCache {
       Some(parent) => self.ensure_dir_exists(parent),
       None => Ok(()),
     }?;
-    fs_util::atomic_write_file(&path, data, crate::http_cache::CACHE_PERM)
+    atomic_write_file(&path, data, crate::http_cache::CACHE_PERM)
       .map_err(|e| with_io_context(&e, format!("{:#?}", &path)))
   }
 }
