+++ cli/http_cache.rs
@@ -3,8 +3,8 @@
 //! as defined in RFC 7234 (<https://tools.ietf.org/html/rfc7234>).
 //! Currently it's a very simplified version to fulfill Deno needs
 //! at hand.
-use crate::fs_util;
 use crate::http_util::HeadersMap;
+use crate::util;
 use deno_core::error::generic_error;
 use deno_core::error::AnyError;
 use deno_core::serde::Deserialize;
@@ -68,7 +68,7 @@ pub fn url_to_filename(url: &Url) -> Option<PathBuf> {
   // NOTE: fragment is omitted on purpose - it's not taken into
   // account when caching - it denotes parts of webpage, which
   // in case of static resources doesn't make much sense
-  let hashed_filename = crate::checksum::gen(&[rest_str.as_bytes()]);
+  let hashed_filename = util::checksum::gen(&[rest_str.as_bytes()]);
   cache_filename.push(hashed_filename);
   Some(cache_filename)
 }
@@ -85,7 +85,7 @@ impl Metadata {
   pub fn write(&self, cache_filename: &Path) -> Result<(), AnyError> {
     let metadata_filename = Self::filename(cache_filename);
     let json = serde_json::to_string_pretty(self)?;
-    fs_util::atomic_write_file(&metadata_filename, json, CACHE_PERM)?;
+    util::fs::atomic_write_file(&metadata_filename, json, CACHE_PERM)?;
     Ok(())
   }
 
@@ -172,7 +172,7 @@ impl HttpCache {
       .expect("Cache filename should have a parent dir");
     self.ensure_dir_exists(parent_filename)?;
     // Cache content
-    fs_util::atomic_write_file(&cache_filename, content, CACHE_PERM)?;
+    util::fs::atomic_write_file(&cache_filename, content, CACHE_PERM)?;
 
     let metadata = Metadata {
       now: SystemTime::now(),
