+++ cli/lsp/completions.rs
@@ -7,9 +7,9 @@ use super::lsp_custom;
 use super::registries::ModuleRegistry;
 use super::tsc;
 
-use crate::fs_util::is_supported_ext;
-use crate::fs_util::relative_specifier;
-use crate::fs_util::specifier_to_file_path;
+use crate::util::path::is_supported_ext;
+use crate::util::path::relative_specifier;
+use crate::util::path::specifier_to_file_path;
 
 use deno_ast::LineAndColumnIndex;
 use deno_ast::SourceTextInfo;
