+++ cli/lsp/registries.rs
@@ -12,8 +12,8 @@ use super::path_to_regex::StringOrNumber;
 use super::path_to_regex::StringOrVec;
 use super::path_to_regex::Token;
 
+use crate::args::CacheSetting;
 use crate::cache::DenoDir;
-use crate::file_fetcher::CacheSetting;
 use crate::file_fetcher::FileFetcher;
 use crate::http_cache::HttpCache;
 use crate::http_util::HttpClient;
