+++ cli/lsp/testing/execution.rs
@@ -6,7 +6,6 @@ use super::lsp_custom;
 
 use crate::args::flags_from_vec;
 use crate::args::DenoSubcommand;
-use crate::checksum;
 use crate::lsp::client::Client;
 use crate::lsp::client::TestingNotification;
 use crate::lsp::config;
@@ -15,6 +14,7 @@ use crate::ops;
 use crate::proc_state;
 use crate::tools::test;
 use crate::tools::test::TestEventSender;
+use crate::util::checksum;
 use crate::worker::create_main_worker_for_test_or_bench;
 
 use deno_core::anyhow::anyhow;
