+++ cli/napi/sym/README.md
