+++ cli/napi/sym/lib.rs
