+++ cli/napi/sym/symbol_exports.json
