+++ cli/tools/standalone.rs
@@ -6,9 +6,9 @@ use crate::args::Flags;
 use crate::args::RunFlags;
 use crate::args::TypeCheckMode;
 use crate::cache::DenoDir;
-use crate::fs_util;
 use crate::standalone::Metadata;
 use crate::standalone::MAGIC_TRAILER;
+use crate::util::path::path_has_trailing_slash;
 use crate::ProcState;
 use deno_core::anyhow::bail;
 use deno_core::anyhow::Context;
@@ -299,7 +299,7 @@ pub fn resolve_compile_executable_output_path(
 ) -> Result<PathBuf, AnyError> {
   let module_specifier = resolve_url_or_path(&compile_flags.source_file)?;
   compile_flags.output.as_ref().and_then(|output| {
-    if fs_util::path_has_trailing_slash(output) {
+    if path_has_trailing_slash(output) {
       let infer_file_name = infer_name_from_url(&module_specifier).map(PathBuf::from)?;
       Some(output.join(infer_file_name))
     } else {
