+++ cli/util/checksum.rs
