+++ cli/util/diff.rs
