+++ cli/util/display.rs
