+++ cli/util/mod.rs
@@ -1,0 +1,13 @@
+// Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
+
+// Note: Only add code in here that has no application specific logic
+pub mod checksum;
+pub mod diff;
+pub mod display;
+pub mod file_watcher;
+pub mod fs;
+pub mod path;
+pub mod progress_bar;
+pub mod text_encoding;
+pub mod unix;
+pub mod windows;
