+++ tools/napi/generate_symbols_lists.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env -S deno run --unstable --allow-read --allow-write
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 
-import exports from "../../cli/napi_sym/symbol_exports.json" assert {
+import exports from "../../cli/napi/sym/symbol_exports.json" assert {
   type: "json",
 };
